add java interop client
Java consumers of the API can not directly interact with suspend functions.
Therefor every java consumer of the API would need to creat their own wrapper.
While possible it might be hard for people with no prior kotlin experience to do so.
To make it easier to have a Java API out of the box I added a new JavaInteropMPSRemoteClient.kt
The client wraps all three suspend operation from the MPSRemoteClient.kt
and exposes them as completable futures that a java programmer can consume.
This introduces a new dependency for the jvm artifact, but I think its worth it.
The alternative would be to introduce another artifact that contains a single class
which seems like an overkill to me.