implement MPSRemoteNode equals
Implement MPSRemoteNode equals via node refrences euqls. That way
we can compare two nodes by accessing if the point to the same
node insid of MPS via its model and node id.

modelix/api-gen runtime will create multiple Java class instances
for the same node depending on how it is used. This way we can
ensure that two INodes pointing to the same MPS node are considered
equal even if they are not the same java object.

As part of the change the dependecies to model-client and model api
were upgraded. Since they have a public API change the minor version
was incremented on this project as well.